Cookbook Start
Added the cookbook object and UI. The player can open and close the cookbook and when they walk away the book closes. I still have to add the other recipes and the ability to flip through the pages, though I've put down the groundwork for that. Also removed old stuff from the level.